Feature: Added song list
Added song list accessible through the settings panel. This allows users
to view their current songs. The new song list modal also allows users
to play and pause songs from this view.

Had to re-do the way we store the songs. Added unique keys which point
to their individual song JSON object. This way we can pair an HTML
element with an id to a song object which contains the location of the
song file.

Songs are now only loaded upon import. This allows songs to be played
instantly when skipping or when selecting a different song through the
song list modal.